#18 refactor: create package for retrofit interface
