+++ halu5071/app/src/main/java/io/moatwel/github/data/datasource/CloudUserDataSource.kt
@@ -22,7 +22,7 @@
 
 package io.moatwel.github.data.datasource
 
-import io.moatwel.github.data.network.UserApi
+import io.moatwel.github.data.network.retrofit.UserApi
 import io.moatwel.github.domain.entity.User
 import io.moatwel.github.presentation.util.setGitHubToken
 import io.reactivex.Observable
