--- halu5071/app/src/main/java/io/moatwel/github/data/network/EventApi.kt
@@ -1,3 +1,0 @@
-package io.moatwel.github.data.network
-
-interface EventApi
