+++ halu5071/app/src/main/java/io/moatwel/github/data/network/retrofit/EventApi.kt
@@ -1,0 +1,3 @@
+package io.moatwel.github.data.network.retrofit
+
+interface EventApi
