+++ halu5071/app/src/main/java/io/moatwel/github/presentation/di/DataSourceModule.kt
@@ -28,7 +28,7 @@ import dagger.Module
 import dagger.Provides
 import io.moatwel.github.data.datasource.AuthDataDataSource
 import io.moatwel.github.data.datasource.CloudUserDataSource
-import io.moatwel.github.data.network.UserApi
+import io.moatwel.github.data.network.retrofit.UserApi
 import okhttp3.OkHttpClient
 
 @Module
