+++ halu5071/app/src/main/java/io/moatwel/github/presentation/di/NetworkModule.kt
@@ -28,7 +28,7 @@ import com.squareup.moshi.KotlinJsonAdapterFactory
 import com.squareup.moshi.Moshi
 import dagger.Module
 import dagger.Provides
-import io.moatwel.github.data.network.UserApi
+import io.moatwel.github.data.network.retrofit.UserApi
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
